fix: moved seeding folder to API project. feat: Added and configured default-categories.json feat: Added newtonsoft json package with test in SeedingService
