+++ src/BudgetManager/BudgetManager.API/BudgetManager.API.csproj
@@ -10,12 +10,25 @@
 	<CodeAnalysisRuleSet>..\StyleCop.ruleset</CodeAnalysisRuleSet>
 	<GenerateDocumentationFile>true</GenerateDocumentationFile>
   </PropertyGroup>
+
+  <ItemGroup>
+    <Content Remove="Seeding\Seeds\country-by-currency-code.json" />
+    <Content Remove="Seeding\Seeds\default-categories.json" />
+    <Content Remove="Seeding\Seeds\default-subcategories.json" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Include="Seeding\Seeds\country-by-currency-code.json" />
+    <None Include="Seeding\Seeds\default-categories.json" />
+  </ItemGroup>
 	
   <ItemGroup>
     <PackageReference Include="AutoMapper" Version="11.0.1" />
     <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="11.0.0" />
     <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="10.0.1" />
     <PackageReference Include="Microsoft.AspNet.WebApi.Cors" Version="5.2.9" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Include="Newtonsoft.Json.Bson" Version="1.0.2" />
     <PackageReference Include="StyleCop.Analyzers" Version="1.1.118">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
