+++ src/BudgetManager/BudgetManager.Model/BudgetManager.Model.csproj
@@ -13,6 +13,7 @@
 	
   <ItemGroup>
     <PackageReference Include="MongoDB.Bson" Version="2.16.1" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
     <PackageReference Include="StyleCop.Analyzers" Version="1.1.118">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
