+++ src/BudgetManager/BudgetManager.Model/DefaultCategory.cs
@@ -1,7 +1,9 @@
 ﻿using System.ComponentModel.DataAnnotations;
+using System.Text.Json.Serialization;
 using BudgetManager.Model.Enums;
 using MongoDB.Bson.Serialization.Attributes;
 using MongoDbGenericRepository.Attributes;
+using Newtonsoft.Json.Converters;
 
 namespace BudgetManager.Model
 {
@@ -12,8 +14,9 @@ namespace BudgetManager.Model
         public Guid Id { get; set; }
         [Required]
         public string Name { get; set; }
-        public List<Guid>? SubCategories { get; set; }
+        public List<DefaultCategory>? SubCategories { get; set; }
         [Required]
+        [JsonConverter(typeof(StringEnumConverter))]
         public OperationType CategoryType { get; set; }
         [Required]
         public string Color { get; set; }
